% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R, R/query.R
\name{bnjr_discover}
\alias{bnjr_discover}
\alias{bjr_discover}
\alias{bjr_query}
\title{Browse available services}
\usage{
bnjr_discover(scan_time = 10L)

bjr_discover(scan_time = 10L)

bjr_query(query, scan_time = 10L)
}
\arguments{
\item{scan_time}{how long to scan for services; default is 10 and
should not really be that much lower in most networks.}
}
\description{
Browse available services
}
