% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{bnjr_query}
\alias{bnjr_query}
\alias{bjr_query}
\alias{mdns_query}
\title{Look for a particular service}
\usage{
bnjr_query(query, scan_time = 10L)

bjr_query(query, scan_time = 10L)

mdns_query(query, scan_time = 10L)
}
\arguments{
\item{query}{service to look for}

\item{scan_time}{how long to scan for services; default is 10 and
should not really be that much lower in most networks.}
}
\value{
data frame
}
\description{
Look for a particular service
}
