% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{bnjr_discover}
\alias{bnjr_discover}
\alias{bjr_discover}
\alias{mdns_discover}
\title{Browse available services}
\usage{
bnjr_discover(scan_time = 10L)

bjr_discover(scan_time = 10L)

mdns_discover(scan_time = 10L)
}
\arguments{
\item{scan_time}{how long to scan for services; default is 10 and
should not really be that much lower in most networks.}
}
\value{
data frame
}
\description{
Browse available services
}
